make locktimes uint64 again

+++ vms/platformvm/txs/add_delegator_test.go
@@ -65,7 +65,7 @@ func TestAddDelegatorTxSyntacticVerify(t *testing.T) {
 	stakes := []*avax.TransferableOutput{{
 		Asset: avax.Asset{ID: ctx.AVAXAssetID},
 		Out: &stakeable.LockOut{
-			Locktime: clk.Time().Add(time.Second).Unix(),
+			Locktime: uint64(clk.Time().Add(time.Second).Unix()),
 			TransferableOut: &secp256k1fx.TransferOutput{
 				Amt: validatorWeight,
 				OutputOwners: secp256k1fx.OutputOwners{
@@ -164,7 +164,7 @@ func TestAddDelegatorTxSyntacticVerifyNotAVAX(t *testing.T) {
 	stakes := []*avax.TransferableOutput{{
 		Asset: avax.Asset{ID: assetID},
 		Out: &stakeable.LockOut{
-			Locktime: clk.Time().Add(time.Second).Unix(),
+			Locktime: uint64(clk.Time().Add(time.Second).Unix()),
 			TransferableOut: &secp256k1fx.TransferOutput{
 				Amt: validatorWeight,
 				OutputOwners: secp256k1fx.OutputOwners{
